pyproject.toml
[tool.poetry]
name = "ersultanchik"
version = "0.1.0"
description = ""
authors = ["pikoUsername <galymzhan.amantaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
django-stubs = "^5.1.3"
django-guardian = "^2.4.0"
django-prometheus = "^2.3.1"
pillow = "^11.1.0"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
